+++ svg.js.d.ts
@@ -88,8 +88,6 @@ declare module '@svgdotjs/svg.js' {
 
   function getWindow(): Window
   function registerWindow(win: Window, doc: Document): void
-  function restoreWindow(): void
-  function saveWindow(): void
   function withWindow(
     win: Window,
     fn: (win: Window, doc: Document) => void